% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigate.R
\name{.copy_to_clipboard}
\alias{.copy_to_clipboard}
\title{Internal function to copy text to clipboard}
\usage{
.copy_to_clipboard(text)
}
\arguments{
\item{text}{Text to copy}
}
\description{
Internal function to copy text to clipboard
}
\keyword{internal}
