% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_text_lines}
\alias{extract_text_lines}
\title{Simple wrapper to extract just the text lines}
\usage{
extract_text_lines(pdf_path)
}
\arguments{
\item{pdf_path}{Path to the PDF file}
}
\value{
Character vector with text lines
}
\description{
Simple wrapper to extract just the text lines
}
